requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.48
opencv-python==4.13.0.92
pytesseract==0.3.13
Pillow==12.1.1
numpy==2.4.3
PyMuPDF==1.27.2.2
reportlab==4.4.10
pytest==9.0.3
scikit-learn==1.7.2
joblib==1.5.2
